requirements.txt
PyQt5
reportlab
pyinstaller
PyQtChart
bcrypt
pyjwt
python-dotenv
openpyxl
Pillow